+++ src/components/App/App.js
@@ -3,7 +3,20 @@ import logo from './logo.svg';
 import './App.css';
 import BusinessList from '../Business List/BusinessList'
 import SearchBar from '../Search Bar/SearchBar'
-import { render } from '@testing-library/react';
+
+var business = {
+  imageSrc: 'https://s3.amazonaws.com/codecademy-content/programs/react/ravenous/pizza.jpg',
+  name: 'Lucas pizza',
+  address: '1010 Paddington Way',
+  city: 'Flavortown',
+  state: 'NY',
+  zipCode: '10101',
+  category: 'Italian',
+  rating: 4.5,
+  reviewCount: 90
+}
+
+const businesses = [business, business, business, business, business, business]
 
 class App extends Component {
   render() {
@@ -11,7 +24,7 @@ class App extends Component {
       <div class="App">
           <h1>ravenous</h1>
           <SearchBar />
-          <BusinessList /> 
+          <BusinessList businesses={businesses}/> 
       </div>
    )
   }
