+++ src/components/Business List/BusinessList.js
@@ -6,12 +6,11 @@ class BusinessList extends Component{
     render(){
         return(
             <div className="BusinessList">
-                <Business />
-                <Business />
-                <Business />
-                <Business />
-                <Business />
-                <Business />
+                {
+                    this.props.businesses.map((business) =>{
+                        return <Business business={business}/>
+                    })
+                }
             </div>
         )
     }
