+++ src/components/Business/Business.js
@@ -1,26 +1,15 @@
-import React, {Component} from 'react'
+import React, { Component } from 'react'
 import './Business.css'
 
 class Business extends Component{
     render() {
-        var business = {
-            imageSrc: 'https://s3.amazonaws.com/codecademy-content/programs/react/ravenous/pizza.jpg',
-            name: 'MarginOtto Pizzeria',
-            address: '1010 Paddington Way',
-            city: 'Flavortown',
-            state: 'NY',
-            zipCode: '10101',
-            category: 'Italian',
-            rating: 4.5,
-            reviewCount: 90
-        }
-
+    const { business } = this.props
     return(
         <div className="Business">
                 <div className="image-container">
                 <img src='https://s3.amazonaws.com/codecademy-content/programs/react/ravenous/pizza.jpg' alt=''/>
                 </div>
-            <h2>MarginOtto Pizzeria</h2>
+            <h2>{business.name}</h2>
             <div className="Business-information">
                 <div className="Business-address">
                 <p>{business.address}</p>
